Moved application state & removed unused
