+++ SoWeQuiches/ApplicationState.swift
@@ -1,0 +1,18 @@
+//
+//  ApplicationState.swift
+//  SoWeQuiches
+//
+//  Created by Zakarya TOLBA on 17/02/2022.
+//
+
+import Foundation
+
+class ApplicationState: ObservableObject {
+    static var shared: ApplicationState = .init()
+    
+    @Published var state: State = .unauthenticated
+    
+    enum State {
+        case authenticated, unauthenticated
+    }
+}
